+++ src/main/java/cc/bitbank/example/Example.java
@@ -21,7 +21,6 @@ import java.util.ResourceBundle;
 public class Example {
     public static void main(String args[]) {
         ResourceBundle rb = ResourceBundle.getBundle("example");
-
         Bitbankcc bb = new Bitbankcc();
         bb.setKey(rb.getString("key"), rb.getString("secret"));
 
@@ -45,7 +44,7 @@ public class Example {
             Order order = bb.getOrder(CurrencyPair.BTC_JPY, 90956209);
             System.out.println(order);
 
-            Order order2 = bb.sendOrder(CurrencyPair.BTC_JPY, 10000, BigDecimal.valueOf(0.01), OrderSide.BUY, OrderType.LIMIT);
+            Order order2 = bb.sendOrder(CurrencyPair.BTC_JPY, BigDecimal.valueOf(10000), BigDecimal.valueOf(0.01), OrderSide.BUY, OrderType.LIMIT);
             System.out.println(order2);
 
             Order order3 = bb.cancelOrder(CurrencyPair.BTC_JPY, 129781978);
@@ -78,7 +77,6 @@ public class Example {
             Withdraw w = bb.requestWithdraw("btc", "XXXXXXX-XXXX-XXXX-XXXX-XXXXXXXXX",
                     BigDecimal.valueOf(0.005), "867005", "");
             System.out.println(w);
-
 
         } catch (BitbankException e) {
             System.out.println(e.code);
